Dateutils monkey patch fix for sub-second string parsing
The kube_config check of the gcp auth expiration fails when running on
a VM due to trying to cast subsecond decimals (a float) directly to an
int from a string.  This must not be a problem on local machines if
I'm the only one running into it.  I monkey patched their datetime
string parsing function as a fix for now, but this should be submitted
as an issue/PR to kubernetes.