added support for bind-mounting local files in run_docker (#49)
Fixes #48. Bind-mounts at the same level as specified by the config dict (directories if absolute path, filename if asset dict).

A side-effect of this is that calling run_docker will trigger an access request on the configuration dictionary. If the user does not have access, the command will fail (not really an issue, since it would also fail inside the docker container in that case). This did require mocking out the access in some other unit tests.

No documentation change needed as the lack of support for bind-mounting local files was undocumented. I did shift one documentation section at the suggestion of @oliverwm1 to make the order of information smoother.

Tests that the correct locations are bind-mounted are added.

Commit history:

* added support for bind-mounting local files in run_docker

* Include another possible gcsfs exception in mock

* remove print statement

* add additional tests, allow single asset to be given

* make reformat

* allow patch_files to be a non-list