Only use --oversubscribe flag if cpu count requires it (#53)
This PR ensures the --oversubscribe MPI flag is only used if the CPU count requires it. It adds the --use-hwthread-cpus flag so that hyperthreaded cores (such as on google VMs and MacBooks) will be used.

It also fixes a bug in _create_job_object by ensuring that the container name is alphanumeric.