Optionally specify runfile with the environmental variable
This allows a user to run the model with a python runfile already present in the
a docker image by setting the FV3_PYTHON_RUNFILE environmental variable. Because
the runfile is in many ways inseperable from the fortran code itself, it is
beneficial to bundle it with the other FV3 code.