Add btl_vader_single_copy_mechanism MPI flag to fv3run (#73)
This PR adds the flag `--mca btl_vader_single_copy_mechanism none` to mpirun in fv3run. It prevents a certain "error" message from being printed repeatedly when the model runs. The error does not affect output, and currently appears so often it is difficult to read the log file.

Commit history:

* add MPI flags which prevent useless error message

* make reformat

* update test with new flag

* update history with note about new flag