Make output capturing an optional behavior
In general, it is more flexible to pipe the outputs to stderr and stdout
and let the user leverage the power of linux shell, pipes, and commands
like tee to put the outputs where they want. Moroever, most standard
kubernetes and docker logging patterns rely on the data appearing in
stdout and stderr.

As an example

    fv3run fv3config.yml output 2> output/stderr.log > output/stdout.log

will emulate the previous behavior provided the output directory is made
first.

That said, this PR leaves the output capturing as an option.