Add an asset type for raw bytes
Previously this library could not deploy arbitrary assets to remote file
systems without first storing those assets in a supported filesystem.

This PR adds a "bytes" asset type that can save an arbitrary list of
bytes to a specified path. Used in conjuction with patch_files, this
will allow the configuration dict to be in principle a
complete/serializable representation of a run-directory.

Motivation:

For example, to nudge the model to observations the namelist requires a
file `nudging_file_list` containing a list of all the nudging files,
this file is not static and is typically generated automatically
depending on the length of the run. Similarly, the diag_table and
runfile for python-wrapped runs must somehow be accessible in the
execution environment.

We have used a variety of methods to transfer these non-fv3config
supported assets into remote execution environments:
- first upload to GCS
- K8s configmaps
- strings to argo workflows which are concatenated to file.
Each of these has its downside, and leads to a lot of duplicated code.
Moreover, each of these methods requires some kind of authentication.