Add DiagTable class (#125)
It is useful to have a richer representation of the diag_table item in fv3config configuration dictionaries.

This PR adds a DiagTable class and associated DiagFileConfig and DiagFieldConfig dataclasses. The DiagTable class has __repr__ and from_str methods which allow going between the Fortran-expected string representation of the diag_table and the python representation. It also includes asdict and from_dict methods, which are useful for serializing/deserializing the diag_table item to/from YAML.

dacite is added as a requirement, as it provides useful functionality for going from a dict to a DataClass with type checking.

This PR does not introduce any breaking changes. If a user provides a path for the diag_table, that file will just be copied to the rundir as is currently done.