Remove MockGCSFS object
This object used some private apis of MemoryFileSystem that have changes.
Specifically, the _strip_protocol. I tried to fix it for a while, but ultimately
concluded that the class would have to be completely reimplemented and tested
throroughly. This is a lot of work for a simple mock...especially given that
MemoryFileSystem already does what we need.

I switched the tests to use URLs starting with "memory://" and deleted the gcsfs
mock. I think this gives us the coverage we need of the protocol handling.