Update testing environment to more recent versions (#146)
* Add minimial example of fsspec>=2022 failure

tox allows easier testing accross versions

the problem is caused by handling of the gs:// prefix by _strip_protocol

* Remove MockGCSFS object

This object used some private apis of MemoryFileSystem that have changes.
Specifically, the _strip_protocol. I tried to fix it for a while, but ultimately
concluded that the class would have to be completely reimplemented and tested
throroughly. This is a lot of work for a simple mock...especially given that
MemoryFileSystem already does what we need.

I switched the tests to use URLs starting with "memory://" and deleted the gcsfs
mock. I think this gives us the coverage we need of the protocol handling.

* allow caching paths /

urls like memory:///dfd did not work before. The gcsfs memory class in
the new version adds a beginning slash for some reason.

* Delete some redundant tests

these tests broke because they hardcoded the cache filename. These tests are
also mostly redundant since the cache filenames are already tested by
test_cache_filename.

Keep the coverage of do_remote_cache.

Remove the "caching caches" test, which seemed a bit tautological and not worth
the maintenance burden.

* update to more recent versions for testing

* Simplify the tox file

* Fix linter

* remove extras from constraints.txt

* remove unused import

* use less confusing name in cache code

* fix tests