Fix `diag_table` bug preventing reproducible restarts (#144)
This PR introduces a `get_time_configuration` function in place of the existing `get_current_date` function.  The `get_time_configuration` function returns both the initialization date and the current date.

The initialization and current dates can be found in lines two and three of any `coupler.res` restart file, and are read in accordingly when one is present in a set of restart files provided as an initial condition.  If the initial condition does not contain a `coupler.res` file, or `coupler_nml.force_date_from_namelist` is set to true, the initialization date and current date are set to the value of the `coupler_nml.current_date` namelist parameter.  

As found in #147, for reproducible restarts, it is critical that the base date written in the diagnostics table is the initialization date of the run (therefore it will remain constant across segments).  Previously in segmented runs, we reset the base date to the date of the current segment, which prevented reproducible restarts.  This PR fixes this issue (see [this notebook](https://github.com/ai2cm/explore/blob/master/spencerc/2022-04-06-restart-reprodicibility/2022-04-07-reproducible-restart.ipynb)).

Resolves #147
See also: ai2cm/fv3gfs-fortran#269