Make fv3config.config_to_asset_list public API (#153)
I would like to enable wandb (and other metadata tracking) integrations for assets supported by fv3config. The "asset list" is the natural target for this, but is not currently a complete representation of the config dict since there is lingering logic in the `write_run_directory` function.

* push all config-handling logic into `config_to_asset_list`
* make `config_to_asset_list` public API

This will allow usages like this downstream:
```
assets = fv3config.config_to_asset_list(config)
for asset in assets:
    if asset["source_location"].startswith("wandb://"):
        artifact = run.use_artifact(...asset)
        path = artifact.download()
        asset["source_location"] = path
    fv3config.write_asset(asset, "rundir")
```

To reviewers: I recommend reviewing commit-by-commit. I tried to isolate the refactors into coherent commits, and it should be much easier to follow that way.